checkin_jump/requirements-dev.txt
pytest==8.3.3
pytest-django==4.9.0
coverage==7.6.1
